requirements.txt
alabaster==0.7.12
ansible==3.1.0
ansible-base==2.10.7
appnope==0.1.0
asgiref==3.3.1
asn1crypto==1.2.0
astroid==2.3.2
astropy==3.0.2
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.1
bitarray==1.0.1
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==0.12.16
boto==2.49.0
Bottleneck==1.2.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
conda==4.7.12
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.7.0
datashape==0.5.4
decorator==4.4.1
defusedxml==0.6.0
distributed==2.7.0
Django==3.1.7
django-debug-toolbar==3.2
django-filter==2.4.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-twilio==0.13.1.post0
djangorestframework==3.12.4
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.9.1
Flask==1.1.1
Flask-Cors==3.0.8
future==0.18.2
geojson==2.5.0
gevent==1.4.0
glob2==0.7
gmaps==0.9.0
gmpy2==2.0.8
greenlet==0.4.15
gunicorn==20.0.4
h5py==2.7.1
HeapDict==1.0.1
heroku==0.1.4
html5lib==1.0.1
idna==2.8
imageio==2.3.0
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
llvmlite==0.30.0
locket==0.2.0
lxml==4.3.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.2
numba==0.38.0
numexpr==2.6.5
numpy==1.14.3
numpydoc==0.9.1
odo==0.5.1
olefile==0.46
openpyxl==2.6.3
packaging==19.2
pandas==0.23.0
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.0
pbr==5.4.3
pep8==1.7.1
pexpect==4.7.0
phonenumbers==8.12.21
pickleshare==0.7.5
Pillow==5.1.0
pkginfo==1.5.0.1
pluggy==0.11.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psutil==5.6.5
psycopg2==2.8.6
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.3
pyodbc==4.0.24
pyOpenSSL==19.0.0
pyparsing==2.4.4
pyrsistent==0.15.4
PySocks==1.7.1
pytest==4.5.0
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==0.5.2
PyYAML==5.1.2
pyzmq==17.1.2
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
raven==6.10.0
requests==2.22.0
rope==0.14.0
ruamel-yaml==0.15.46
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
snowballstemmer==2.0.0
sortedcollections==1.1.1
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.10
sqlparse==0.4.1
statsmodels==0.9.0
stevedore==1.31.0
sympy==1.4
tables==3.4.3
TBB==0.1
tblib==1.5.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
twilio==6.56.0
typed-ast==1.4.0
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.24.2
virtualenv==16.7.5
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.2
xlwings==0.16.0
xlwt==1.2.0
zict==1.0.0
zipp==0.6.0
